fix: invites are not longer being deleted when a new invite is createdt

+++ src/modules/sharing/sharing.service.ts
@@ -1178,10 +1178,12 @@ export class SharingService {
     itemId: Sharing['itemId'],
     type: Sharing['type'],
   ) {
-    await this.sharingRepository.deleteInvitesBy({
-      itemId,
-      itemType,
-    });
+    if (type === SharingType.Private) {
+      await this.sharingRepository.deleteInvitesBy({
+        itemId,
+        itemType,
+      });
+    }
 
     await this.sharingRepository.deleteSharingsBy({
       itemId,
